macropad: fix phantom scroll from unsigned wrap-around compare
The encoder scrolled continuously in one direction while untouched.

MT6701_ANGLE_MAX was defined as (1U << 14), making it and the derived
MT6701_ANGLE_HALF unsigned int. In `delta > MT6701_ANGLE_HALF` and
`delta < -MT6701_ANGLE_HALF` the usual arithmetic conversions therefore
promoted the signed delta to unsigned, and -MT6701_ANGLE_HALF evaluated
to 4294959104U. The second branch matched every small value including
zero, so a stationary sensor added a full 16384-count rotation on every
poll tick -- 200 phantom rotations per second at the 5ms poll rate,
always in the same direction.

Make both constants signed and widen delta to int32_t so the wrap
correction only fires on a real 0/16383 boundary crossing. At rest delta
is now 0 and nothing is reported; +/-1 LSB jitter cancels in the
microdegree accumulator.

+++ drivers/sensor/mt6701/mt6701.c
@@ -19,7 +19,10 @@ LOG_MODULE_DECLARE(zmk, CONFIG_ZMK_LOG_LEVEL);
 
 #define MT6701_ANGLE_REG 0x03
 #define MT6701_ANGLE_BITS 14
-#define MT6701_ANGLE_MAX (1U << MT6701_ANGLE_BITS)
+/* Deliberately signed: these are compared against a signed delta below. As unsigned
+ * they would drag the delta through the usual arithmetic conversions and make the
+ * wrap correction fire on every value, including zero. */
+#define MT6701_ANGLE_MAX (1 << MT6701_ANGLE_BITS)
 #define MT6701_ANGLE_HALF (MT6701_ANGLE_MAX / 2)
 #define MT6701_MICRODEG_PER_ROTATION 360000000LL
 #define MICRODEG_PER_DEGREE 1000000LL
@@ -68,7 +71,7 @@ static bool mt6701_accumulate(const struct device *dev) {
         return false;
     }
 
-    int16_t delta = raw - data->prev_angle;
+    int32_t delta = (int32_t)raw - (int32_t)data->prev_angle;
     if (delta > MT6701_ANGLE_HALF) {
         delta -= MT6701_ANGLE_MAX;
     } else if (delta < -MT6701_ANGLE_HALF) {
